# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: API de gestion des articles
  description: Documentation de l'API pour la gestion des articles.
  version: 1.0.0
servers:
  - url: /api
    description: Serveur de développement local
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Article:
      type: object
      properties:
        id:
          type: integer
        libelle:
          type: string
        description:
          type: string
        price:
          type: number
        stock:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
paths:
...
  /v1/articles/{id}:
    delete:
      summary: Supprimer un article par ID
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Article supprimé avec succès
  /v1/articles/update-stock:
    post:
      summary: Mettre à jour le stock d'articles
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Stock mis à jour avec succès
  /v1/articles/{id}/quantity:
    patch:
      summary: Mettre à jour la quantité d'un article
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Quantité mise à jour avec succès
  /v1/articles:
    get:
      summary: Obtenir les articles disponibles
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Liste des articles disponibles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Article"
...